feat(crush-lu): Comprehensive improvement - security, performance, code org, frontend quality
Phase 1 - Security:
- Use secrets.compare_digest() for constant-time token comparison
- Add rate limiting to event registration, connection requests, messaging
- Fix 4 silent exception blocks to log warnings

Phase 2 - Database Performance:
- Fix N+1 query in coach_journey_dashboard using annotate()
- Add 5 indexes on EventConnection and ProfileSubmission
- Add custom QuerySet/Manager methods (for_user, pending_for_user, etc.)

Phase 3 - Code Organization:
- Split views.py (5,018 lines) into 8 focused modules
- Re-export pattern keeps urls.py unchanged

Phase 4 - @coach_required Decorator:
- New decorator checks active coach and attaches request.coach
- Replaces 16 duplicate try/except blocks across coach views

Phase 5 - Frontend Quality:
- Fix Alpine CSP violations in photo_slideshow and email_template_manager
- Remove ~88 console.log statements from production JS
- Add loading="lazy" to 6 below-fold images
- Add ARIA labels to 14 icon-only buttons
- Fix missing alt text on profile photo
