+++ crush_lu/static/crush_lu/js/sw-unregister.js
@@ -8,11 +8,7 @@
     if ('serviceWorker' in navigator) {
         navigator.serviceWorker.getRegistrations().then(function(registrations) {
             registrations.forEach(function(registration) {
-                registration.unregister().then(function(success) {
-                    if (success) {
-                        console.log('[DEV] Service Worker unregistered for development');
-                    }
-                });
+                registration.unregister();
             });
         });
     }
