+++ crush_lu/templates/crush_lu/coach_review_profile.html
@@ -37,13 +37,13 @@
                 {# Photos - Compact View #}
                 <div class="flex shrink-0 gap-2">
                     {% if profile.photo_1 %}
-                        <img src="{{ profile.photo_1.url }}" class="w-20 h-20 shrink-0 rounded-lg object-cover border-2 border-white dark:border-gray-600 shadow-md dark:shadow-gray-900/20" alt="Photo 1">
+                        <img src="{{ profile.photo_1.url }}" class="w-20 h-20 shrink-0 rounded-lg object-cover border-2 border-white dark:border-gray-600 shadow-md dark:shadow-gray-900/20" alt="Photo 1" loading="lazy">
                     {% endif %}
                     {% if profile.photo_2 %}
-                        <img src="{{ profile.photo_2.url }}" class="w-20 h-20 shrink-0 rounded-lg object-cover border-2 border-white dark:border-gray-600 shadow-md dark:shadow-gray-900/20" alt="Photo 2">
+                        <img src="{{ profile.photo_2.url }}" class="w-20 h-20 shrink-0 rounded-lg object-cover border-2 border-white dark:border-gray-600 shadow-md dark:shadow-gray-900/20" alt="Photo 2" loading="lazy">
                     {% endif %}
                     {% if profile.photo_3 %}
-                        <img src="{{ profile.photo_3.url }}" class="w-20 h-20 shrink-0 rounded-lg object-cover border-2 border-white dark:border-gray-600 shadow-md dark:shadow-gray-900/20" alt="Photo 3">
+                        <img src="{{ profile.photo_3.url }}" class="w-20 h-20 shrink-0 rounded-lg object-cover border-2 border-white dark:border-gray-600 shadow-md dark:shadow-gray-900/20" alt="Photo 3" loading="lazy">
                     {% endif %}
                 </div>
 
@@ -230,7 +230,7 @@
                 </svg>
                 {% trans "Review Guidelines" %}
             </h3>
-            <button @click="close" class="text-white hover:text-gray-200 transition-colors">
+            <button @click="close" class="text-white hover:text-gray-200 transition-colors" aria-label="{% trans 'Close review guidelines' %}">
                 <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                     <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"/>
                 </svg>
@@ -291,6 +291,7 @@
             x-cloak
             class="fixed bottom-6 right-6 bg-gradient-to-r from-purple-600 to-pink-500 text-white p-3 rounded-full shadow-lg hover:shadow-xl transition-shadow"
             title="{% trans 'Show Review Guidelines' %}"
+            aria-label="{% trans 'Show Review Guidelines' %}"
         >
             <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9.663 17h4.673M12 3v1m6.364 1.636l-.707.707M21 12h-1M4 12H3m3.343-5.657l-.707-.707m2.828 9.9a5 5 0 117.072 0l-.548.547A3.374 3.374 0 0014 18.469V19a2 2 0 11-4 0v-.531c0-.895-.356-1.754-.988-2.386l-.548-.547z"/>
