+++ crush_lu/templates/crush_lu/create_profile.html
@@ -103,7 +103,8 @@
                  class="bg-red-50 border border-red-200 text-red-700 px-4 py-3 rounded-lg mb-4 relative">
                 {{ form.non_field_errors }}
                 <button type="button" @click="dismiss"
-                        class="absolute top-2 right-2 text-red-500 hover:text-red-700">
+                        class="absolute top-2 right-2 text-red-500 hover:text-red-700"
+                        aria-label="{% trans 'Dismiss' %}">
                     <svg class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20">
                         <path fill-rule="evenodd" d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z" clip-rule="evenodd"/>
                     </svg>
@@ -125,7 +126,8 @@
                     {% endfor %}
                 </ul>
                 <button type="button" @click="dismiss"
-                        class="absolute top-2 right-2 text-red-500 hover:text-red-700">
+                        class="absolute top-2 right-2 text-red-500 hover:text-red-700"
+                        aria-label="{% trans 'Dismiss' %}">
                     <svg class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20">
                         <path fill-rule="evenodd" d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z" clip-rule="evenodd"/>
                     </svg>
@@ -534,7 +536,7 @@
                         <path fill-rule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clip-rule="evenodd"/>
                     </svg>
                     <span x-text="saveError"></span>
-                    <button type="button" @click="clearSaveError" class="ml-auto text-red-500 hover:text-red-700">
+                    <button type="button" @click="clearSaveError" class="ml-auto text-red-500 hover:text-red-700" aria-label="{% trans 'Dismiss error' %}">
                         <svg class="w-4 h-4" fill="currentColor" viewBox="0 0 20 20">
                             <path fill-rule="evenodd" d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z" clip-rule="evenodd"/>
                         </svg>
@@ -759,7 +761,7 @@
                         <path fill-rule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clip-rule="evenodd"/>
                     </svg>
                     <span x-text="saveError"></span>
-                    <button type="button" @click="clearSaveError" class="ml-auto text-red-500 hover:text-red-700">
+                    <button type="button" @click="clearSaveError" class="ml-auto text-red-500 hover:text-red-700" aria-label="{% trans 'Dismiss error' %}">
                         <svg class="w-4 h-4" fill="currentColor" viewBox="0 0 20 20">
                             <path fill-rule="evenodd" d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z" clip-rule="evenodd"/>
                         </svg>
@@ -809,7 +811,8 @@
                         </label>
                         <button type="button" x-show="photo1HasImage" @click="removePhoto1"
                                 class="absolute -top-2 -right-2 w-6 h-6 bg-red-500 text-white rounded-full
-                                       flex items-center justify-center hover:bg-red-600 transition shadow-lg">
+                                       flex items-center justify-center hover:bg-red-600 transition shadow-lg"
+                                aria-label="{% trans 'Remove photo' %} 1">
                             <svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"/>
                             </svg>
@@ -836,7 +839,8 @@
                         </label>
                         <button type="button" x-show="photo2HasImage" @click="removePhoto2"
                                 class="absolute -top-2 -right-2 w-6 h-6 bg-red-500 text-white rounded-full
-                                       flex items-center justify-center hover:bg-red-600 transition shadow-lg">
+                                       flex items-center justify-center hover:bg-red-600 transition shadow-lg"
+                                aria-label="{% trans 'Remove photo' %} 2">
                             <svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"/>
                             </svg>
@@ -863,7 +867,8 @@
                         </label>
                         <button type="button" x-show="photo3HasImage" @click="removePhoto3"
                                 class="absolute -top-2 -right-2 w-6 h-6 bg-red-500 text-white rounded-full
-                                       flex items-center justify-center hover:bg-red-600 transition shadow-lg">
+                                       flex items-center justify-center hover:bg-red-600 transition shadow-lg"
+                                aria-label="{% trans 'Remove photo' %} 3">
                             <svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"/>
                             </svg>
@@ -1011,7 +1016,7 @@
                         <path fill-rule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clip-rule="evenodd"/>
                     </svg>
                     <span x-text="saveError"></span>
-                    <button type="button" @click="clearSaveError" class="ml-auto text-red-500 hover:text-red-700">
+                    <button type="button" @click="clearSaveError" class="ml-auto text-red-500 hover:text-red-700" aria-label="{% trans 'Dismiss error' %}">
                         <svg class="w-4 h-4" fill="currentColor" viewBox="0 0 20 20">
                             <path fill-rule="evenodd" d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z" clip-rule="evenodd"/>
                         </svg>
@@ -1190,7 +1195,7 @@
                  class="relative bg-white dark:bg-gray-800 rounded-2xl shadow-xl max-w-md w-full p-6">
 
                 <!-- Close button -->
-                <button @click="close" class="absolute top-4 right-4 text-gray-400 hover:text-gray-600">
+                <button @click="close" class="absolute top-4 right-4 text-gray-400 hover:text-gray-600" aria-label="{% trans 'Close' %}">
                     <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"/>
                     </svg>
