+++ crush_lu/templates/crush_lu/partials/_push_activation_prompt.html
@@ -45,7 +45,8 @@
 
         <!-- Close button -->
         <button @click="dismissPermanently"
-                class="flex-shrink-0 text-white/60 hover:text-white">
+                class="flex-shrink-0 text-white/60 hover:text-white"
+                aria-label="{% trans 'Dismiss' %}">
             <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"/>
             </svg>
