+++ crush_lu/templates/crush_lu/partials/pwa_install_banner.html
@@ -60,7 +60,7 @@
         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"/>
     </svg>
     <span class="text-green-800 dark:text-green-300 font-medium text-sm">{% trans "Crush.lu installed! Access it from your home screen." %}</span>
-    <button type="button" @click="close" class="ml-2 text-green-400 dark:text-green-500 hover:text-green-600 dark:hover:text-green-300 transition-colors">
+    <button type="button" @click="close" class="ml-2 text-green-400 dark:text-green-500 hover:text-green-600 dark:hover:text-green-300 transition-colors" aria-label="{% trans 'Dismiss' %}">
         <svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"/>
         </svg>
