+++ crush_lu/templates/crush_lu/profile_submitted_luxid_mockup.html
@@ -302,7 +302,8 @@
                             <div class="flex-shrink-0">
                                 <img src="{{ submission.coach.photo.url }}"
                                      alt="{{ submission.coach.user.get_full_name }}"
-                                     class="w-16 h-16 rounded-full object-cover border-2 border-crush-purple/30 shadow-md">
+                                     class="w-16 h-16 rounded-full object-cover border-2 border-crush-purple/30 shadow-md"
+                                     loading="lazy">
                             </div>
                             {% else %}
                             <div class="flex-shrink-0">
